Move unquantize to astropy/io/fits/_tiled_compression as it isn't exposed in the CFITSIO library and because it is edited from imcompress.c so doesn't make sense to keep it in cextern
